+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+* The _class info cache_ is now initialized silently by default. This results in less confusing output.
+  * You can now `@orchard.java/cache-initializer` for deterministically waiting for this cache workload to complete.
+  * You can control its verbosity by setting `"-Dorchard.initialize-cache.silent=false"` (or `[...]=true`).
+
 ## 0.6.5 (2021-02-13)
 
 ### Bugs Fixed